+++ packages/console/src/Console/Artisan.php
@@ -41,6 +41,7 @@ final class Artisan extends SymfonyApplication
             new \Fx\Framework\Console\Commands\PackageInstallCommand($this->root, true),
             new AboutCommand($this->root), new OptimizeClearCommand(),
             new \Fx\Framework\Console\Commands\SetupInitCommand(),
+            new \Fx\Framework\Console\Commands\SetupUpgradeCommand($this->root),
             new LegacyCommand('make:controller', 'Cria um controller', true),
             new LegacyCommand('cache:clear', 'Limpa o cache da aplicacao'),
             new LegacyCommand('cache:cleartmp', 'Limpa arquivos temporarios'),